Fix using sent-mail, drafts and templates folders
These functions from kdepimlibs using some checker if folder is akonadi
specified and if not, then returning some default values. This causing
that it is not possible to use for example Sent folder on imap reource.

 * KPIMIdentities::Identity::setFcc()
 * KPIMIdentities::Identity::setDrafts()
 * KPIMIdentities::Identity::setTemplates()
 * KPIMIdentities::Identity::fcc()
 * KPIMIdentities::Identity::drafts()
 * KPIMIdentities::Identity::templates()

kdepimlibs storing these folders to config file emailidentities in kde
home folder in configs "Fcc", "Drafts" and "Templates". To coexistent
with new akonadi kmail, store it to "Fcc2", "Drafts2" and "Templates2".